backend/pyproject.toml
[tool.poetry]
name = "paradiso"
version = "1.2"
description = "Ecommerce WebApp"
authors = ["Ajas Mohammed <ajasmohammed33@gmail.com>"]
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
build = "1.0.3"
CacheControl = "0.13.1"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
cryptography = "42.0.1"
defusedxml = "0.7.1"
distlib = "0.3.8"
Django = "5.0.1"
django-allauth = "0.60.1"
django-cors-headers = "4.3.1"
django-otp = "1.3.0"
django-phonenumber-field = "7.3.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
dulwich = "0.21.7"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
idna = "3.6"
importlib-metadata = "7.0.1"
installer = "0.7.0"
# jaraco.classes = "3.3.1"
keyring = "24.3.0"
more-itertools = "10.2.0"
msgpack = "1.0.7"
oauthlib = "3.2.2"
packaging = "23.2"
pexpect = "4.9.0"
pillow = "10.2.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pycparser = "2.21"
PyJWT = "2.8.0"
pyproject_hooks = "1.0.0"
python-decouple = "3.8"
python3-openid = "3.2.0"
pytz = "2023.4"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.6.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
six = "1.16.0"
sqlparse = "0.4.4"
stripe = "8.0.0"
tomlkit = "0.12.3"
trove-classifiers = "2024.1.31"
typing_extensions = "4.9.0"
tzdata = "2023.4"
urllib3 = "2.1.0"
virtualenv = "20.25.0"
zipp = "3.17.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
